added in takePicture icon into camera UI
no functionality yet, just a UI update.
Can change, but I was thinking that since we probably need to standardise our train and test data, would also be good to standardise a large circle in the middle of the camera UI, so they have a guide to make sure that their whole 'plate' is within view.