Complete item/store overhaul
This update makes it much easier to add in items (background lol) - functionality for display items (equip pet items) still needs to be added, but not in this sprint.

Basically created global app functions for purchasing items and storing the store items in an array to add in stuff easier.